[Feature] Support flow control in zeta (#309)

+++ seatunnel-api/src/main/java/org/apache/seatunnel/api/table/type/SeaTunnelRow.java
@@ -34,7 +34,7 @@ public final class SeaTunnelRow implements Serializable {
     /** The array to store the actual internal format values. */
     private final Object[] fields;
 
-    private volatile long size;
+    private volatile int size;
 
     public SeaTunnelRow(int arity) {
         this.fields = new Object[arity];
@@ -102,7 +102,7 @@ public final class SeaTunnelRow implements Serializable {
 
     public long getBytesSize(SeaTunnelRowType rowType) {
         if (size == 0) {
-            long s = 0;
+            int s = 0;
             for (int i = 0; i < fields.length; i++) {
                 s += getBytesForValue(fields[i], rowType.getFieldType(i));
             }
@@ -194,9 +194,9 @@ public final class SeaTunnelRow implements Serializable {
         }
     }
 
-    public long getBytesSize() {
+    public int getBytesSize() {
         if (size == 0) {
-            long s = 0;
+            int s = 0;
             for (int i = 0; i < fields.length; i++) {
                 s += getBytesForValue(fields[i]);
             }
